Fixed vNeueStrecke(), p_dZeit in Fahrzeug
*Changed p_dZeit type from int to double
*Added class Weg to Fahrzeug.h

+++ Aufgabenblock_2/AktivesVO.h
@@ -21,7 +21,7 @@ private:
 protected:
 	string p_sName;
 	int p_iID;
-	int p_dZeit;
+	double p_dZeit;
 
 };
